requirements.txt
Flask
flask-cors
python-dotenv
pyodbc